+++ package.json
@@ -8,7 +8,7 @@
     "deep-equal": "^2.0.4",
     "proskomma": "^0.1.21",
     "react-dom": "^17.0.1",
-    "react-scripts": "4.0.0",
+    "styled-components": "^5.2.1",
     "xregexp": "^4.3.0"
   },
   "scripts": {
@@ -16,12 +16,6 @@
     "start": "styleguidist server",
     "build": "styleguidist build",
     "prepublishOnly": "rm -fr ./dist & babel ./src --out-dir ./dist -s inline --extensions \".ts,.js\""
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
   },
   "browserslist": {
     "production": [
@@ -36,11 +30,14 @@
     ]
   },
   "devDependencies": {
-    "@babel/register": "^7.12.1",
+    "@babel/preset-react": "^7.12.7",
+    "babel-loader": "^8.2.2",
     "babel-tape-runner": "^3.0.0",
     "cheerio": "^1.0.0-rc.3",
+    "css-loader": "^5.0.1",
     "react": "^17.0.1",
     "react-styleguidist": "^11.1.5",
+    "style-loader": "^2.0.0",
     "tap-summary": "^4.0.0",
     "tape": "^5.0.1",
     "webpack": "^5.10.0"
