--- public/manifest.json
@@ -1,9 +1,0 @@
-{
-  "short_name": "Proskomma React App",
-  "name": "Create Proskomma React Demo App",
-  "icons": [],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
