+++ src/components/Button/README.md
@@ -1,0 +1,3 @@
+```js
+<Button />
+```
