--- src/components/Example.css
@@ -1,55 +1,0 @@
-.App {
-    padding: 10px
-}
-
-.heading {
-    padding-bottom: 30px;
-    font-size: xx-large;
-    font-weight: bold;
-    font-style: italic
-}
-
-.heading img {
-    padding-right: 10px
-}
-
-.PkForm {
-    padding-bottom: 20px
-}
-
-.PkFormRow {
-    margin-top: 5px;
-}
-
-.formLabel {
-    font-size: small;
-    display: inline-block;
-    width: 96px;
-    text-align: right;
-    font-weight: bold;
-    padding-right: 10px
-}
-
-.formTextInput {
-    background-color: #EEE
-}
-
-.PkDisplay {
-    font-size: x-small;
-    font-family: monospace;
-    white-space: pre-wrap
-}
-
-.PkDisplay-content {
-    font-size: x-large;
-    font-weight: bold
-}
-
-.PkDisplay-queryTime {
-    margin-top: 20px;
-    font-style: italic;
-}
-
-.PkDisplay-rawResults, .PkDisplay-queries {
-    margin-top: 20px
-}
