+++ src/components/index.js
@@ -1,3 +1,4 @@
 export { default as PkDisplay } from "./PkDisplay";
 export { default as PkForm } from "./PkForm";
 export { default as PkQuery } from "./PkQuery";
+export { default as Button } from './Button'
